requirements.txt
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
python-dotenv==0.19.1
openai==0.27.0
Pillow==8.3.1
numpy==1.21.0
tensorflow==2.5.0
groq